Precompile the reinitialization folder
