+++ CMakeLists.txt
@@ -107,6 +107,8 @@ if(ADAFLO)
   message("-- Adaflo found: ${ADAFLO}")
   target_compile_definitions(${projectName} PUBLIC MELT_POOL_DG_WITH_ADAFLO)
   target_include_directories(${projectName} PUBLIC ${ADAFLO_INCLUDE})
+  target_compile_definitions(${libName} PUBLIC MELT_POOL_DG_WITH_ADAFLO)
+  target_include_directories(${libName} PUBLIC ${ADAFLO_INCLUDE})
   target_link_libraries(${projectName} ${ADAFLO})
   set(DEAL_II_WITH_ADAFLO_SUPPORT ON)
 endif()
